% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawCovidMap.R
\name{drawCovidMap}
\alias{drawCovidMap}
\title{drawCovidMap}
\usage{
drawCovidMap(date, type)
}
\arguments{
\item{date}{the date of which you want the drawn map}

\item{type}{type of the map. Either Deaths or Confirmed cases}
}
\value{
a leaflet map
}
\description{
Draws the global map of daily covid-19 deaths or confirmed cases
}
\examples{
mapOfCovid <- drawCovidMap("07-03-2020","Deaths")

}
