% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawCovidPlot.R
\name{drawCovidPlot}
\alias{drawCovidPlot}
\title{drawCovidPlot}
\usage{
drawCovidPlot(period, country)
}
\arguments{
\item{period}{a vector containing the start and end date of the time series}

\item{country}{name of the chosen country}
}
\value{
a ggplot object
}
\description{
Draws the plot of the number of covid-19 deaths and confirmed cases during a time series in a specific country
}
\examples{
plotOfCovid <- drawCovidPlot(c("04-02-2020","06-12-2020"),"US")

}
