% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateCovid.R
\name{updateCovid}
\alias{updateCovid}
\title{updateCovid}
\usage{
updateCovid()
}
\value{
a dataframe of data
}
\description{
Gets updated covid-19 data from Johns Hopkins University github
}
\examples{
updatedCovidData <- updateCovid()

}
