% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateCovid.R
\name{updateCovid}
\alias{updateCovid}
\title{updateCovid}
\usage{
updateCovid(date)
}
\arguments{
\item{date}{the date to which you want to update your data}
}
\value{
a dataframe of data
}
\description{
Gives updated covid-19 data from Johns Hopkins University's github
}
\examples{
updatedCovidData <- updateCovid()

}
